+++ lib/include/stdio.h
@@ -9,3 +9,4 @@ int putchar(int c);
 int puts(char *s);
 int sprintf(char *buf, char *fmt, ...);
 int printf(char *fmt, ...);
+int sscanf(char *str, char *fmt, ...);